- int overflow for average size of serialized objects fixed - recalculate negative Page memory size with new average size of serialized objects

+++ h2/src/main/org/h2/mvstore/type/ObjectDataType.java
@@ -1530,7 +1530,7 @@ public class ObjectDataType implements DataType {
             int size = data.length * 2;
             // adjust the average size
             // using an exponential moving average
-            averageSize = (size + 15 * averageSize) / 16;
+            averageSize = (int) ((size + 15L * averageSize) / 16);
             buff.put((byte) TYPE_SERIALIZED_OBJECT).putVarInt(data.length)
                     .put(data);
         }
@@ -1542,7 +1542,7 @@ public class ObjectDataType implements DataType {
             int size = data.length * 2;
             // adjust the average size
             // using an exponential moving average
-            averageSize = (size + 15 * averageSize) / 16;
+            averageSize = (int) ((size + 15L * averageSize) / 16);
             buff.get(data);
             return deserialize(data);
         }
